New system for effects, switching between them, and presentation demo
AudioConnections are not meant to be changed on the fly, so the best way
to switch between effects is to have one "switcher" effect that is
always connected. This effect will let you choose between its list of
effects.

The effects themselves can then be converted to a simpler object, which
is now InterfaceEffect. They subclass from this. All these objects have
to do is edit blocks of samples, and the rest of the routing details are
taken care of by the switcher.

Along with this is a demo which